backend/requirements.txt
 
fastapi
uvicorn
requests
pydantic
sqlalchemy
reportlab
dnspython
python-multipart
